Add some digest and crypto algorithms
1. Add MD5 and SHA digest algorithms
2. Add AES crypto algorithm
3. Add ukive::net::HttpClient (incomplete)
4. Add ukive::net::tls (incomplete)
5. Add ukive::BigInteger series for searching primes

+++ ukive/drawable/check_drawable.cpp
@@ -25,8 +25,8 @@ namespace ukive {
 
         float hp_offset = 0.f;
         float total_offset = 0.f;
-        int length = std::ceil(win_->dpToPx(16));
-        int stroke_width = win_->dpToPx(2.5f);
+        int length = int(std::ceil(win_->dpToPx(16)));
+        int stroke_width = int(win_->dpToPx(2.5f));
         if (stroke_width % 2) {
             hp_offset = 0.5f;
         }
